requirements.txt
fastapi
uvicorn
langgraph
langgraph-checkpoint-sqlite
httpx
pydantic
langchain-ollama
langchain-experimental
beautifulsoup4
requests
docker
python-multipart
